+++ src/modules/nfts/components/wizard/Edition.tsx
@@ -1,4 +1,4 @@
-import NavContainer from '@/common/components/wizard/NavContainer';
+import NavContainer from '@/modules/nfts/components/wizard/NavContainer';
 import { Divider, Form, FormInstance, Space, Row } from 'antd';
 import React from 'react';
 import Image from 'next/image';
