+++ src/modules/nfts/components/wizard/PriceSummary.tsx
@@ -1,16 +1,14 @@
-import NavContainer from '@/common/components/wizard/NavContainer';
-import { Divider, Form, FormInstance, Row, Col } from 'antd';
+import { Divider, Form, Row, Col } from 'antd';
 import React from 'react';
 import Image from 'next/image';
 import { StepWizardChildProps } from 'react-step-wizard';
 import styled from 'styled-components';
 import Button from '@/common/components/elements/Button';
 import Paragraph from 'antd/lib/typography/Paragraph';
+import NavContainer from '@/modules/nfts/components/wizard/NavContainer';
 
 interface Props extends Partial<StepWizardChildProps> {
   images: Array<File>;
-  index: number;
-  form: FormInstance;
 }
 
 const Grid = styled.div`
@@ -43,7 +41,7 @@ const PriceRow = () => {
   );
 };
 
-export default function PriceSummary({ previousStep, goToStep, images, nextStep, form }: Props) {
+export default function PriceSummary({ previousStep, goToStep, images, nextStep }: Props) {
   const handleNext = () => {
     nextStep!();
   };
